pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "neural-lattice"
version = "0.1.0"
description = "Neural Lattice Cognitive Architecture — zone-based document management with cognitive load tracking"
requires-python = ">=3.11"
license = {text = "CC-BY-4.0"}
dependencies = [
    "fastapi>=0.110",
    "uvicorn[standard]>=0.27",
    "pyyaml>=6.0",
    "pydantic>=2.5",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=4.1",
    "httpx>=0.27",
    "ruff>=0.3",
]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
